[#29] Add `mapMaybe`
Resolves #29
