+++ src/Slist/Size.hs
@@ -1,8 +1,10 @@
 {- |
 Copyright:  (c) 2019-2020 Veronika Romashkina
             (c) 2020-2021 Kowainik
-License:    MPL-2.0
+SPDX-License-Identifier: MPL-2.0
 Maintainer: Kowainik <xrom.xkov@gmail.com>
+Stability:   Stable
+Portability: Portable
 
 Lists size representation.
 -}
